+++ package.json
@@ -22,7 +22,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "react-scripts start npx eslint '**/*.{js,jsx}' --fix"
   },
   "eslintConfig": {
     "extends": [
