+++ src/components/singlebook/DeleteBook.jsx
@@ -1,8 +1,8 @@
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { deleteBook, fetchBooks } from '../redux/books/booksSlice';
+import { deleteBook, fetchBooks } from '../../redux/books/booksSlice';
 
-const DeleteBookButton = ({ id }) => {
+const DeleteBook = ({ id }) => {
   const dispatch = useDispatch();
 
   const handleDelete = () => {
@@ -21,8 +21,8 @@ const DeleteBookButton = ({ id }) => {
   );
 };
 
-DeleteBookButton.propTypes = {
+DeleteBook.propTypes = {
   id: PropTypes.string.isRequired,
 };
 
-export default DeleteBookButton;
+export default DeleteBook;
